requirements.txt
# OpenClaw Multi-Agent GameDev System
# 核心依赖
pyyaml>=6.0

# Web仪表盘
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 开发工具（可选）
pytest>=7.4.0
